chore(css): move CSSHelper class in a file and export from index.ts
